TX shuffle
Lint

[WIP] wip

[ghstack-poisoned]
